+++ app/templates/profile.html
@@ -316,7 +316,7 @@
             {# small toggle so the owner can flip privacy on/off without leaving
                the page. Strip is green when public, amber when private; the icon
                flips between an open eye and a struck-through eye. #}
-            <form method="POST" action="{{ url_for('profile_toggle_following_privacy') }}"
+            <form method="POST" action="{{ url_for('users.profile_toggle_following_privacy') }}"
                   class="profile-following-privacy {{ 'is-public' if user.following_list_public else 'is-private' }}">
                 <span class="profile-following-status">
                     {% if user.following_list_public %}
@@ -349,11 +349,11 @@
             <ul class="profile-following-list">
                 {% for f in following_users %}
                     <li class="profile-following-item">
-                        <a class="profile-following-avatar" href="{{ url_for('user_profile_page', username=f.username) }}" aria-hidden="true">
+                        <a class="profile-following-avatar" href="{{ url_for('users.user_profile_page', username=f.username) }}" aria-hidden="true">
                             {{ f.username[0]|upper }}
                         </a>
                         <div class="profile-following-info">
-                            <a class="profile-following-name" href="{{ url_for('user_profile_page', username=f.username) }}">{{ f.username }}</a>
+                            <a class="profile-following-name" href="{{ url_for('users.user_profile_page', username=f.username) }}">{{ f.username }}</a>
                             <span class="profile-following-handle">@{{ f.username|lower }}</span>
                         </div>
                     </li>
@@ -372,7 +372,7 @@
         {% set can_view_followers = is_own_profile or user.followers_list_public %}
 
         {% if is_own_profile %}
-            <form method="POST" action="{{ url_for('profile_toggle_followers_privacy') }}"
+            <form method="POST" action="{{ url_for('users.profile_toggle_followers_privacy') }}"
                   class="profile-following-privacy {{ 'is-public' if user.followers_list_public else 'is-private' }}">
                 <span class="profile-following-status">
                     {% if user.followers_list_public %}
@@ -405,11 +405,11 @@
             <ul class="profile-following-list">
                 {% for f in follower_users %}
                     <li class="profile-following-item">
-                        <a class="profile-following-avatar" href="{{ url_for('user_profile_page', username=f.username) }}" aria-hidden="true">
+                        <a class="profile-following-avatar" href="{{ url_for('users.user_profile_page', username=f.username) }}" aria-hidden="true">
                             {{ f.username[0]|upper }}
                         </a>
                         <div class="profile-following-info">
-                            <a class="profile-following-name" href="{{ url_for('user_profile_page', username=f.username) }}">{{ f.username }}</a>
+                            <a class="profile-following-name" href="{{ url_for('users.user_profile_page', username=f.username) }}">{{ f.username }}</a>
                             <span class="profile-following-handle">@{{ f.username|lower }}</span>
                         </div>
                     </li>
