+++ app/templates/profile_edit.html
@@ -7,7 +7,7 @@
 {% block content %}
 <section class="settings-page">
     <header class="settings-header">
-        <a href="{{ url_for('profile_page') }}" class="profile-edit-back">← Back to my profile</a>
+        <a href="{{ url_for(users.profile_page') }}" class="profile-edit-back">← Back to my profile</a>
         <h1 class="settings-title">Edit profile</h1>
         <p class="settings-subtitle">Change how you appear to other users. Account &amp; security settings live on the Settings page.</p>
     </header>
